+++ tests/testthat/tests.make_stancode.R
@@ -824,24 +824,6 @@ test_that("fixed residual covariance matrices appear in the Stan code", {
                "target += multi_student_t_lpdf(Y | nu, mu, V)")
 })
 
-test_that("Stan code for BSTS models is correct", {
-  dat <- data.frame(y = rnorm(10), x = rnorm(10))
-  scode <- SW(make_stancode(
-    y ~ x, data = dat, autocor = cor_bsts(),
-    prior = prior(normal(0, 10), sigmaLL)
-  ))
-  expect_match2(scode, "mu[n] += loclev[n]")
-  expect_match2(scode, "target += normal_lpdf(loclev[n] | loclev[n - 1], sigmaLL)")
-  expect_match2(scode, "target += normal_lpdf(sigmaLL | 0, 10)")
-  
-  dat <- data.frame(y = rexp(1), x = rnorm(10))
-  scode <- SW(make_stancode(
-    y~x, data = dat, family = student("log"),
-    autocor = cor_bsts()
-  ))
-  expect_match2(scode, "target += normal_lpdf(loclev[n] | log(Y[n]), sigmaLL)")
-})
-
 test_that("Stan code for GAMMs is correct", {
   dat <- data.frame(y = rnorm(10), x = rnorm(10), g = factor(rep(1:2, 5)))
   scode <- make_stancode(y ~ s(x) + (1|g), data = dat,